move solvers into dedicated folder
