+++ biosnicar/rt_solvers/__init__.py
@@ -1,0 +1,11 @@
+"""Radiative transfer solvers for biosnicar.
+
+This module contains implementations of different radiative transfer solvers:
+- Toon matrix method solver
+- Adding-doubling solver
+"""
+
+from biosnicar.rt_solvers.toon_rt_solver import toon_solver
+from biosnicar.rt_solvers.adding_doubling_solver import adding_doubling_solver
+
+__all__ = ["toon_solver", "adding_doubling_solver"]
