+++ biosnicar/rt_solvers/adding_doubling_solver.py
@@ -32,7 +32,7 @@ solid ice layers and fresnel reflection are included.
 import numpy as np
 from scipy.signal import savgol_filter
 
-from biosnicar.classes import Outputs
+from biosnicar.classes.outputs import Outputs
 
 
 def adding_doubling_solver(tau, ssa, g, L_snw, ice, illumination, model_config):
