+++ biosnicar/rt_solvers/toon_rt_solver.py
@@ -28,7 +28,7 @@ radiative transfer models.
 import numpy as np
 from scipy.signal import savgol_filter
 
-from biosnicar.classes import Outputs
+from biosnicar.classes.outputs import Outputs
 
 
 def toon_solver(tau, ssa, g, L_snw, ice, illumination, model_config, rt_config):
